Rename package to reflect project scope
