+++ src/test/java/org/cru/cas/client/integration/CopyCasAttributesToHeadersFilterTest.java
@@ -1,6 +1,6 @@
-package org.cru.userheader;
+package org.cru.cas.client.integration;
 
-import static org.cru.userheader.CopyCasAttributesToHeadersFilter.ATTRIBUTE_MAPPING_PARAMETER;
+import static org.cru.cas.client.integration.CopyCasAttributesToHeadersFilter.ATTRIBUTE_MAPPING_PARAMETER;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -11,12 +11,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Enumeration;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
